# onebrain-web/src/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "ServiceKeyInfo": {
        "properties": {
          "account_id": {
            "default": "",
            "title": "Account Id",
            "type": "string"
          },
          "app_id": {
            "default": "",
            "title": "App Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "default": "",
            "title": "Label",
            "type": "string"
          },
          "last_used_at": {
            "default": "",
            "title": "Last Used At",
            "type": "string"
          },
          "last_used_endpoint": {
            "default": "",
            "title": "Last Used Endpoint",
            "type": "string"
          },
          "purposes": {
            "items": {
              "type": "string"
            },
            "title": "Purposes",
            "type": "array"
          },
          "revoked_at": {
            "default": "",
            "title": "Revoked At",
            "type": "string"
          },
          "rotated_from_id": {
            "default": "",
            "title": "Rotated From Id",
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "space_ids": {
            "items": {
              "type": "string"
            },
            "title": "Space Ids",
            "type": "array"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "use_count": {
            "default": 0,
            "title": "Use Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "scopes"
        ],
        "title": "ServiceKeyInfo",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "onebrain",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/operator/accounts/{account_id}/service-keys": {
      "get": {
        "operationId": "list_account_service_keys_api_operator_accounts__account_id__service_keys_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServiceKeyInfo"
                  },
                  "title": "Response List Account Service Keys Api Operator Accounts  Account Id  Service Keys Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Account Service Keys",
        "tags": [
          "operator"
        ]
      }
    },
    "/api/operator/accounts/{account_id}/service-keys/{key_id}": {
      "delete": {
        "operationId": "revoke_account_service_key_api_operator_accounts__account_id__service_keys__key_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Account Service Key",
        "tags": [
          "operator"
        ]
      }
    },
    "/api/operator/customers": {
      "get": {
        "operationId": "list_customers_api_operator_customers_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CustomerOverviewOut"
                  },
                  "title": "Response List Customers Api Operator Customers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Customers",
        "tags": [
          "operator"
        ]
      }
    },
    "/api/operator/deployments": {
      "get": {
        "operationId": "list_deployments_api_operator_deployments_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeploymentOut"
                  },
                  "title": "Response List Deployments Api Operator Deployments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Deployments",
        "tags": [
          "operator"
        ]
      },
      "post": {
        "operationId": "create_deployment_api_operator_deployments_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Deployment",
        "tags": [
          "operator"
        ]
      }
    },
    "/api/operator/deployments/{deployment_id}/backups": {
      "post": {
        "operationId": "record_backup_api_operator_deployments__deployment_id__backups_post",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Backup",
        "tags": [
          "operator"
        ]
      }
    },
    "/api/operator/deployments/{deployment_id}/backups/latest": {
      "get": {
        "operationId": "latest_backup_api_operator_deployments__deployment_id__backups_latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "ob_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ob Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BackupOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Latest Backup Api Operator Deployments  Deployment Id  Backups Latest Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Backup",
        "tags": [
          "operator"
        ]
      }
    },
...
  }
}